Added a return construct to the language
A program can optionally add a 'return x,y,z...' instruction as a last
command. This will specify the output variables, to be used with
output sensitive typing (and ignored in the simple typechecking algo).
This change affected the abstract syntax and thus every modules.

Added a few test cases.